SM: Moved modal configuration to seperate file.
